move extensions to a gltf folder, some updates
